requirements.txt
requests
docker
PyYaml
click
colorama